add shuffle and refactor
