+++ module-2-case-study.iml
@@ -7,6 +7,6 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="jid3lib-0.5.41" level="project" />
+    <orderEntry type="library" name="jid3lib-0.5.4" level="project" />
   </component>
 </module>
