+++ src/model/Mp3Collection.java
@@ -1,4 +1,4 @@
-package song;
+package model;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
