+++ src/model/Mp3Parser.java
@@ -1,4 +1,4 @@
-package song;
+package model;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
