+++ src/model/Mp3Song.java
@@ -1,4 +1,4 @@
-package song;
+package model;
 
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
